Fix unlinking test file with file still open
The script used a `with open(...` to open the write test file and it also unlinked and returned inside the scope of the with. On some systems Python holding the file open will prevent unlinking and cause an exception, so the script will erroneously assume it is unable to write to the target directory even though it does have the correct permissions.

This commit moves the unlink and return outside the scope of the with statement that opens the file, which will cause Python to close the file before unlinking it.